Add files via upload
•	Life expectancy – Regression Analysis
This project aimed to explore the difference of best predictors for life expectancy in developed and developing countries. In this project, we collected nearly 200 countries' life expectancy data and trained them with linear regression. In order to optimize the linear regression model and verify the interchangeability of the linear model for developed, developing countries and worldwide, we utilized the best regression subsets and model accuracy metrics, such as Cp, AIC, and adjusted R^2. We have found that the best predictor for developing and developed countries is very different. Thinness is the best predictor for developed countries, and Schooling, HIV and AIDS are the most influential factors for developing countries. We also found that the linear model for developing countries is near to the worldwide linear model.